% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.gcPCA}
\alias{coef.gcPCA}
\alias{coef.sparse_gcPCA}
\title{Extract gcPCA loadings}
\usage{
\method{coef}{gcPCA}(object, ...)

\method{coef}{sparse_gcPCA}(object, ...)
}
\arguments{
\item{object}{A fitted \code{"gcPCA"} or \code{"sparse_gcPCA"} object.}

\item{...}{Unused.}
}
\value{
For dense models, a loadings matrix. For sparse models, a named list
of loading matrices (one per lasso penalty).
}
\description{
Extract gcPCA loadings
}
