% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.gcPCA}
\alias{fitted.gcPCA}
\title{Return fitted training scores}
\usage{
\method{fitted}{gcPCA}(object, ...)
}
\arguments{
\item{object}{A fitted \code{"gcPCA"} or \code{"sparse_gcPCA"} object.}

\item{...}{Passed to \code{predict()}.}
}
\value{
A list with \code{Ra_scores} and \code{Rb_scores}.
}
\description{
Return fitted training scores
}
