% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{gcpca-package}
\alias{gcpca}
\alias{gcpca-package}
\title{gcpca package}
\description{
Dense and sparse generalized contrastive principal component analysis.
}
\seealso{
\code{\link[=gcPCA]{gcPCA()}} and \code{\link[=sparse_gcPCA]{sparse_gcPCA()}}
}
\author{
\strong{Maintainer}: Eliezyer de Oliveira \email{eliezyer.deoliveira@gmail.com}

}
\keyword{internal}
