% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadings}
\alias{loadings}
\title{Get gcPCA loadings}
\usage{
loadings(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Extra arguments passed to \code{coef()}.}
}
\value{
A loading matrix or a named list of loading matrices.
}
\description{
Convenience wrapper around \code{coef()} for \code{"gcPCA"} and \code{"sparse_gcPCA"}
objects.
}
