% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gcPCA}
\alias{plot.gcPCA}
\title{Plot gcPCA scores}
\usage{
\method{plot}{gcPCA}(x, which = c(1, 2), ...)
}
\arguments{
\item{x}{A fitted \code{"gcPCA"} or \code{"sparse_gcPCA"} object.}

\item{which}{Integer vector of length 2 indicating components to plot.}

\item{...}{Extra graphical parameters passed to \code{plot()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot gcPCA scores
}
