% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gcPCA}
\alias{print.gcPCA}
\title{Print a gcPCA model}
\usage{
\method{print}{gcPCA}(x, ...)
}
\arguments{
\item{x}{A fitted \code{"gcPCA"} or \code{"sparse_gcPCA"} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a gcPCA model
}
