% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.gcPCA}
\alias{print.summary.gcPCA}
\title{Print a gcPCA summary}
\usage{
\method{print}{summary.gcPCA}(x, ...)
}
\arguments{
\item{x}{A \code{"summary.gcPCA"} object from \code{\link[=summary.gcPCA]{summary.gcPCA()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a gcPCA summary
}
