% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scores}
\alias{scores}
\title{Get gcPCA scores}
\usage{
scores(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{newdata}{Optional matrix (treated as \code{Ra}) or list with \code{Ra}/\code{Rb}.}

\item{...}{Extra arguments passed to \code{predict()}.}
}
\value{
A list of projected scores.
}
\description{
Convenience wrapper around \code{predict()} for \code{"gcPCA"} and \code{"sparse_gcPCA"}
objects.
}
